SourceApplication/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Bcrypt==1.0.1
PyJWT==2.0.0
kafka-python==2.1.3